pydantic-ai-agent-mcp-tool/pyproject.toml
[project]
name = "pydantic-ai-agent-mcp-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.18.0",
    "pydantic-ai-slim[groq,mcp]>=1.1.0",
    "python-dotenv>=1.1.1",
]
